@@ -11,11 +11,11 @@ import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.block.entity.BlockEntity;
-import net.minecraft.enchantment.ProtectionEnchantment;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.ItemEntity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.TntEntity;
+import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.fluid.FluidState;
 import net.minecraft.item.ItemStack;
@@ -137,17 +137,17 @@ public class BonesburrierExplosion extends CustomExplosion {
             double w;
 
             if (ProtectionsProvider.canDamageEntity(value, damageSource)) {
-                if (value.isImmuneToExplosion() || !((w = Math.sqrt(value.squaredDistanceTo(vec3d)) / (double) q) <= 1.0) || (aa = Math.sqrt((x = value.getX() - this.x) * x + (y = (value instanceof TntEntity ? value.getY() : value.getEyeY()) - this.y) * y + (z = value.getZ() - this.z) * z)) == 0.0)
+                if (value.isImmuneToExplosion(this) || !((w = Math.sqrt(value.squaredDistanceTo(vec3d)) / (double) q) <= 1.0) || (aa = Math.sqrt((x = value.getX() - this.x) * x + (y = (value instanceof TntEntity ? value.getY() : value.getEyeY()) - this.y) * y + (z = value.getZ() - this.z) * z)) == 0.0)
                     continue;
                 x /= aa;
                 y /= aa;
                 z /= aa;
                 double ab = Explosion.getExposure(vec3d, value);
                 double ac = (1.0 - w) * ab;
-                value.damage(this.getDamageSource(), (int) ((ac * ac + ac) / 2.0 * 7.0 * (double) q + 1.0));
+                value.damage(damageSource, (int) ((ac * ac + ac) / 2.0 * 7.0 * (double) q + 1.0));
                 double ad = ac;
-                if (value instanceof LivingEntity) {
-                    ad = ProtectionEnchantment.transformExplosionKnockback((LivingEntity) value, ac);
+                if (value instanceof LivingEntity living) {
+                    ad *= living.getAttributeValue(EntityAttributes.GENERIC_EXPLOSION_KNOCKBACK_RESISTANCE);
                 }
                 value.setVelocity(value.getVelocity().add(x * ad, y * ad, z * ad));
                 if (!(value instanceof PlayerEntity) || (playerEntity = (PlayerEntity) value).isSpectator() || playerEntity.isCreative() && playerEntity.getAbilities().flying)
@@ -163,7 +163,7 @@ public class BonesburrierExplosion extends CustomExplosion {
     public void affectWorld(boolean particles) {
         boolean bl;
         if (this.world.isClient) {
-            this.world.playSound(this.x, this.y, this.z, SoundEvents.ENTITY_GENERIC_EXPLODE, SoundCategory.BLOCKS, 4.0f, (1.0f + (this.world.random.nextFloat() - this.world.random.nextFloat()) * 0.2f) * 0.7f, false);
+            this.world.playSound(this.x, this.y, this.z, SoundEvents.ENTITY_GENERIC_EXPLODE.value(), SoundCategory.BLOCKS, 4.0f, (1.0f + (this.world.random.nextFloat() - this.world.random.nextFloat()) * 0.2f) * 0.7f, false);
         }
         boolean bl2 = bl = this.destructionType != DestructionType.KEEP;
         if (particles) {
