@@ -7,7 +7,7 @@ import net.minecraft.client.render.entity.ProjectileEntityRenderer;
 import net.minecraft.util.Identifier;
 
 public class AmethystShardEntityRenderer extends ProjectileEntityRenderer<AmethystShardEntity> {
-    public static final Identifier TEXTURE = new Identifier(Blast.MODID, "textures/entity/projectiles/amethyst_shard.png");
+    public static final Identifier TEXTURE = Identifier.of(Blast.MODID, "textures/entity/projectiles/amethyst_shard.png");
 
     public AmethystShardEntityRenderer(EntityRendererFactory.Context context) {
         super(context);
