@@ -1,29 +1,25 @@
 package ladysnake.blast.common;
 
-import ladysnake.blast.common.init.BlastBlocks;
-import ladysnake.blast.common.init.BlastEntities;
-import ladysnake.blast.common.init.BlastItems;
-import ladysnake.blast.common.init.BlastSoundEvents;
+import ladysnake.blast.common.init.*;
 import net.fabricmc.api.ModInitializer;
 import net.minecraft.entity.data.TrackedDataHandler;
 import net.minecraft.entity.data.TrackedDataHandlerRegistry;
-import net.minecraft.network.PacketByteBuf;
+import net.minecraft.network.RegistryByteBuf;
+import net.minecraft.network.codec.PacketCodec;
 import net.minecraft.util.math.Direction;
 
 public class Blast implements ModInitializer {
     public static final String MODID = "blast";
 
     public static final TrackedDataHandler<Direction> FACING = new TrackedDataHandler<>() {
-        public void write(PacketByteBuf packetByteBuf, Direction direction) {
-            packetByteBuf.writeEnumConstant(direction);
+        @Override
+        public PacketCodec<? super RegistryByteBuf, Direction> codec() {
+            return Direction.PACKET_CODEC;
         }
 
-        public Direction read(PacketByteBuf packetByteBuf) {
-            return packetByteBuf.readEnumConstant(Direction.class);
-        }
-
-        public Direction copy(Direction direction) {
-            return direction;
+        @Override
+        public Direction copy(Direction value) {
+            return value;
         }
     };
 
