@@ -16,6 +16,7 @@ import net.minecraft.network.packet.s2c.play.GameStateChangeS2CPacket;
 import net.minecraft.particle.ItemStackParticleEffect;
 import net.minecraft.particle.ParticleTypes;
 import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.sound.SoundEvents;
@@ -34,8 +35,8 @@ public class AmethystShardEntity extends PersistentProjectileEntity {
     }
 
     @Override
-    protected ItemStack asItemStack() {
-        return new ItemStack(Items.AIR);
+    protected ItemStack getDefaultItemStack() {
+        return Items.AIR.getDefaultStack();
     }
 
     @Override
@@ -111,9 +112,8 @@ public class AmethystShardEntity extends PersistentProjectileEntity {
             if (entity instanceof LivingEntity) {
                 LivingEntity livingEntity = (LivingEntity) entity;
 
-                if (!this.getWorld().isClient && entity2 instanceof LivingEntity) {
-                    EnchantmentHelper.onUserDamaged(livingEntity, entity2);
-                    EnchantmentHelper.onTargetDamaged((LivingEntity) entity2, livingEntity);
+                if (!this.getWorld().isClient) {
+                    EnchantmentHelper.onTargetDamaged((ServerWorld) getWorld(), livingEntity, damageSource2);
                 }
 
                 this.onHit(livingEntity);
