@@ -12,6 +12,7 @@ import net.minecraft.item.Item;
 import net.minecraft.item.Items;
 import net.minecraft.network.packet.s2c.play.GameStateChangeS2CPacket;
 import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.sound.SoundEvents;
@@ -61,9 +62,8 @@ public class IcicleEntity extends AmethystShardEntity {
             if (isEnderman)
                 return;
             if (entity instanceof LivingEntity livingEntity) {
-                if (!this.getWorld().isClient && entity2 instanceof LivingEntity) {
-                    EnchantmentHelper.onUserDamaged(livingEntity, entity2);
-                    EnchantmentHelper.onTargetDamaged((LivingEntity) entity2, livingEntity);
+                if (!this.getWorld().isClient) {
+                    EnchantmentHelper.onTargetDamaged((ServerWorld) getWorld(), livingEntity, damageSource2);
                 }
 
                 this.onHit(livingEntity);
