@@ -7,7 +7,7 @@ import net.minecraft.sound.SoundEvent;
 import net.minecraft.util.Identifier;
 
 public class BlastSoundEvents {
-    public static final SoundEvent PIPE_BOMB_TICK = SoundEvent.of(new Identifier(Blast.MODID, "entity.pipe_bomb.tick"));
+    public static final SoundEvent PIPE_BOMB_TICK = SoundEvent.of(Identifier.of(Blast.MODID, "entity.pipe_bomb.tick"));
 
     public static void initialize() {
         Registry.register(Registries.SOUND_EVENT, PIPE_BOMB_TICK.getId(), PIPE_BOMB_TICK);
