@@ -3,18 +3,14 @@
  */
 package ladysnake.blast.common.recipe;
 
-import ladysnake.blast.common.init.BlastItems;
-import net.minecraft.inventory.RecipeInputInventory;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
-import net.minecraft.nbt.NbtCompound;
-import net.minecraft.nbt.NbtList;
 import net.minecraft.recipe.Ingredient;
 import net.minecraft.recipe.RecipeSerializer;
 import net.minecraft.recipe.SpecialCraftingRecipe;
 import net.minecraft.recipe.book.CraftingRecipeCategory;
-import net.minecraft.registry.DynamicRegistryManager;
-import net.minecraft.util.Identifier;
+import net.minecraft.recipe.input.CraftingRecipeInput;
+import net.minecraft.registry.RegistryWrapper;
 import net.minecraft.world.World;
 
 public class PipeBombRecipe extends SpecialCraftingRecipe {
@@ -23,16 +19,16 @@ public class PipeBombRecipe extends SpecialCraftingRecipe {
     private static final Ingredient DURATION_MODIFIER = Ingredient.ofItems(Items.GUNPOWDER);
     private static final Ingredient FIREWORK_STAR = Ingredient.ofItems(Items.FIREWORK_STAR);
 
-    public PipeBombRecipe(Identifier identifier) {
-        super(identifier, CraftingRecipeCategory.EQUIPMENT);
+    public PipeBombRecipe() {
+        super(CraftingRecipeCategory.EQUIPMENT);
     }
 
     @Override
-    public boolean matches(RecipeInputInventory craftingInventory, World world) {
+    public boolean matches(CraftingRecipeInput input, World world) {
         boolean bl = false;
         int i = 0;
-        for (int j = 0; j < craftingInventory.size(); ++j) {
-            ItemStack itemStack = craftingInventory.getStack(j);
+        for (int j = 0; j < input.getSize(); ++j) {
+            ItemStack itemStack = input.getStackInSlot(j);
             if (itemStack.isEmpty()) continue;
             if (PAPER.test(itemStack)) {
                 if (bl) {
@@ -48,37 +44,33 @@ public class PipeBombRecipe extends SpecialCraftingRecipe {
     }
 
     @Override
-    public ItemStack craft(RecipeInputInventory craftingInventory, DynamicRegistryManager registryManager) {
+    public ItemStack craft(CraftingRecipeInput input, RegistryWrapper.WrapperLookup lookup) {
         ItemStack itemStack = new ItemStack(Items.FIREWORK_ROCKET, 3);
-        NbtCompound nbtCompound = itemStack.getOrCreateSubNbt("Fireworks");
-        NbtList nbtList = new NbtList();
-        int i = 0;
-        for (int j = 0; j < craftingInventory.size(); ++j) {
-            NbtCompound nbtCompound2;
-            ItemStack itemStack2 = craftingInventory.getStack(j);
-            if (itemStack2.isEmpty()) continue;
-            if (DURATION_MODIFIER.test(itemStack2)) {
-                ++i;
-                continue;
-            }
-            if (!FIREWORK_STAR.test(itemStack2) || (nbtCompound2 = itemStack2.getSubNbt("Explosion")) == null) continue;
-            nbtList.add(nbtCompound2);
-        }
-        nbtCompound.putByte("Flight", (byte) i);
-        if (!nbtList.isEmpty()) {
-            nbtCompound.put("Explosions", nbtList);
-        }
+        // todo fix crafting
+//        NbtCompound nbtCompound = itemStack.getOrCreateSubNbt("Fireworks");
+//        NbtList nbtList = new NbtList();
+//        int i = 0;
+//        for (int j = 0; j < input.getSize(); ++j) {
+//            NbtCompound nbtCompound2;
+//            ItemStack itemStack2 = input.getStackInSlot(j);
+//            if (itemStack2.isEmpty()) continue;
+//            if (DURATION_MODIFIER.test(itemStack2)) {
+//                ++i;
+//                continue;
+//            }
+//            if (!FIREWORK_STAR.test(itemStack2) || (nbtCompound2 = itemStack2.getSubNbt("Explosion")) == null) continue;
+//            nbtList.add(nbtCompound2);
+//        }
+//        nbtCompound.putByte("Flight", (byte) i);
+//        if (!nbtList.isEmpty()) {
+//            nbtCompound.put("Explosions", nbtList);
+//        }
         return itemStack;
     }
 
     @Override
     public boolean fits(int width, int height) {
         return width * height >= 2;
-    }
-
-    @Override
-    public ItemStack getOutput(DynamicRegistryManager registryManager) {
-        return new ItemStack(BlastItems.PIPE_BOMB);
     }
 
     @Override
