@@ -6,7 +6,6 @@ import net.minecraft.inventory.StackReference;
 import net.minecraft.item.BundleItem;
 import net.minecraft.item.ItemStack;
 import net.minecraft.screen.slot.Slot;
-import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.util.ClickType;
 import org.spongepowered.asm.mixin.Mixin;
@@ -20,7 +19,7 @@ public class BundleItemMixin {
     public void onStackClicked(ItemStack stack, Slot slot, ClickType clickType, PlayerEntity player, CallbackInfoReturnable<Boolean> callbackInfoReturnable) {
         ItemStack itemStack = slot.getStack();
         if (itemStack.isOf(BlastItems.PIPE_BOMB)) {
-            player.playSound(SoundEvents.BLOCK_TRIPWIRE_CLICK_ON, SoundCategory.NEUTRAL, 0.5F, 1.0f);
+            player.playSound(SoundEvents.BLOCK_TRIPWIRE_CLICK_ON, 0.5F, 1);
         }
     }
 
@@ -29,7 +28,7 @@ public class BundleItemMixin {
     public void onClicked(ItemStack stack, ItemStack otherStack, Slot slot, ClickType clickType, PlayerEntity player, StackReference cursorStackReference, CallbackInfoReturnable<Boolean> callbackInfoReturnable) {
         ItemStack itemStack = slot.getStack();
         if (itemStack.isOf(BlastItems.PIPE_BOMB)) {
-            player.playSound(SoundEvents.BLOCK_TRIPWIRE_CLICK_ON, SoundCategory.NEUTRAL, 0.5F, 1.0f);
+            player.playSound(SoundEvents.BLOCK_TRIPWIRE_CLICK_ON, 0.5F, 1);
         }
     }
 }
